feat: FilterMut is function now
