+++ filter_test.go
@@ -1,6 +1,10 @@
 package collection
 
-import "testing"
+import (
+	"fmt"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
 
 func TestFilter(t *testing.T) {
 	t.Run("Test FilterAny", func(t *testing.T) {
@@ -58,4 +62,35 @@ func TestFilter(t *testing.T) {
 			}
 		}
 	})
+}
+
+func TestFilterMut(t *testing.T) {
+	t.Run("Test FilterMut - inplace change", func(t *testing.T) {
+		for _, data := range []struct {
+			in           Collection[int]
+			deleteValues []int
+			expect       Collection[int]
+		}{
+			{[]int{1, 2, 3}, []int{1}, []int{2, 3}},
+			{[]int{1, 2, 3}, []int{6}, []int{1, 2, 3}},
+			{[]int{1, 2, 3}, nil, []int{1, 2, 3}},
+			{[]int{1, 2, 3}, []int{77, 99}, []int{1, 2, 3}},
+			{[]int{1, 2, 3}, []int{2}, []int{1, 3}},
+			{[]int{1, 2, 3, 4, 5, 6}, []int{3, 4, 6}, []int{1, 2, 5}},
+		} {
+			ptr := fmt.Sprintf("%p", data.in)
+			nuData := FilterMut(data.in, func(i int) bool {
+				for _, idx := range data.deleteValues {
+					if idx == i {
+						return false
+					}
+				}
+				return true
+			})
+			assert.Equal(t, data.expect, nuData)
+			ptrNu := fmt.Sprintf("%p", nuData)
+			assert.Equal(t, ptr, ptrNu)
+		}
+	})
+
 }
